todoproject/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
Faker==12.0.1
idna==3.3
importlib-metadata==4.11.1
Markdown==3.3.6
mixer==7.2.2
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
zipp==3.7.0
